Extract cronItemMatches to new file
