+++ src/cron.ts
@@ -1,6 +1,7 @@
 import * as assert from "assert";
 import { Pool } from "pg";
 
+import { cronItemMatches } from "./cronMatcher";
 import { parseCrontab } from "./crontab";
 import defer from "./deferred";
 import getCronItems from "./getCronItems";
@@ -561,37 +562,4 @@ function digestTimestamp(ts: Date): TimestampDigest {
   const month = ts.getUTCMonth() + 1;
   const dow = ts.getUTCDay();
   return { min, hour, date, month, dow };
-}
-
-/**
- * Returns true if the cronItem should fire for the given timestamp digest,
- * false otherwise.
- */
-export function cronItemMatches(
-  cronItem: ParsedCronItem,
-  digest: TimestampDigest,
-): boolean {
-  const { min, hour, date, month, dow } = digest;
-
-  if (
-    // If minute, hour and month match
-    cronItem.minutes.includes(min) &&
-    cronItem.hours.includes(hour) &&
-    cronItem.months.includes(month)
-  ) {
-    const dateIsExclusionary = cronItem.dates.length !== 31;
-    const dowIsExclusionary = cronItem.dows.length !== 7;
-    if (dateIsExclusionary && dowIsExclusionary) {
-      // Cron has a special behaviour: if both date and day of week are
-      // exclusionary (i.e. not "*") then a match for *either* passes.
-      return cronItem.dates.includes(date) || cronItem.dows.includes(dow);
-    } else if (dateIsExclusionary) {
-      return cronItem.dates.includes(date);
-    } else if (dowIsExclusionary) {
-      return cronItem.dows.includes(dow);
-    } else {
-      return true;
-    }
-  }
-  return false;
 }
